Extend ReadStream with SetConnClosedCallback
Description:
Removed the propagation of connectionClosed channel when creating
a new ReadStream and instead extended ReadStream with a new function
that allows to set a callback for a closed connection. This looks
cleaner and enables to set the callback outside of constructor which
is important for disconnected publisher handling (next diff).